Actualizado mostrado de cursorFran y comprobacion de que la tabla comparar solo inserta si apunta a Producto y ProductoCompetidor

+++ Ignacio/sql/producto.sql
@@ -6,5 +6,5 @@ CREATE TABLE Producto (
   Especificaciones  VARCHAR(60),
   Piezas VARCHAR(50),
   Stock    INT,
-  Precio    INT
+  Precio   REAL
 );
